Added test path for cmdlet homepage

+++ scripts/RemoveBoilerPlateCode.ps1
@@ -69,9 +69,12 @@ function Repair-Examples{
      $destination = Join-Path $WorkLoadDocsPath $GraphProfilePath $Path
      $CmdletHomePage = Join-Path $destination "$path.md"
      $NoDescriptionAvailable = "{{ Fill in the Description }}"
-     $HomePageContent = Get-Content $CmdletHomePage
-     $HomePageContent = $HomePageContent -replace $NoDescriptionAvailable, ""
-     $HomePageContent | Out-File $CmdletHomePage -Encoding UTF8
+     if(Test-Path $CmdletHomePage){
+        $HomePageContent = Get-Content $CmdletHomePage
+        $HomePageContent = $HomePageContent -replace $NoDescriptionAvailable, ""
+        $HomePageContent | Out-File $CmdletHomePage -Encoding UTF8
+     }
+
      foreach ($File in Get-ChildItem $destination) {
         $NoCodeAvailable = "Add code here"
         $SearchBlock = "## EXAMPLES(?s).*## PARAMETERS"
